[ICU-6991] Code coverage for missing functions in Calendar, DecimalFormat, MessageFormat, PluralFormat, and StringList. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/26613 Former-commit-id: c2cfe3a24a21ea06d4e2d3cd00b26e8f3233799a

+++ icu4c/source/test/intltest/caltest.cpp
@@ -408,7 +408,7 @@ CalendarTest::TestGenericAPI()
     }
 
     Calendar *cal3 = Calendar::createInstance(status);
-    cal3->roll(Calendar::SECOND, 0, status);
+    cal3->roll(Calendar::SECOND, (int32_t)0, status);
     if (failure(status, "Calendar::roll(EDateFields, int32_t, UErrorCode)")) return;
 
     delete cal;
